Yield tasks from `Engine.map()`; move `Task` class to `yapapi.runner.task`.
